Refactor for implementing OTP management with SSL support and recovery code generation and update Makefile
